refactor: dependency decoupling
