+++ Bitkit/ViewModels/Trezor/TrezorViewModel.swift
@@ -1,6 +1,5 @@
 import BitkitCore
 import Combine
-import CoreBluetooth
 import Foundation
 
 /// Represents the current step in the send transaction flow
@@ -59,98 +58,30 @@ enum TrezorAccountTypeSelection: String, CaseIterable, Identifiable, CustomStrin
     }
 }
 
-/// ViewModel for Trezor hardware wallet integration
+/// ViewModel for the Trezor hardware-wallet dev/dashboard screens. Owns only the dev-screen
+/// tools (address generation, signing, lookups, the single dev watcher, etc.); device/connection
+/// orchestration and pairing/PIN/passphrase coordination live in `TrezorManager`, injected here
+/// as `connection`.
 @Observable
 @MainActor
 class TrezorViewModel {
-    // MARK: - Network Configuration
+    // MARK: - Connection Manager
 
-    /// The network selected in the Trezor dashboard (independent of app's global network)
-    var selectedNetwork: TrezorCoinType
+    /// Device/connection orchestration and pairing/PIN/passphrase coordination.
+    let connection: TrezorManager
 
-    /// Map the app's current network to the corresponding TrezorCoinType (used for default initialization)
-    static var appDefaultCoinType: TrezorCoinType {
-        switch Env.network {
-        case .bitcoin: .bitcoin
-        case .testnet: .testnet
-        case .signet: .signet
-        case .regtest: .regtest
-        }
-    }
-
-    /// BIP44 coin type component based on the dashboard's selected network: "0'" for mainnet, "1'" for test networks
-    var coinTypeComponent: String {
-        selectedNetwork == .bitcoin ? "0'" : "1'"
-    }
-
-    /// BIP44 coin type component based on the app's global network (used for initial default values)
-    private static var defaultCoinTypeComponent: String {
-        Env.network == .bitcoin ? "0'" : "1'"
-    }
-
-    // MARK: - Connection State
-
-    /// Whether the Trezor manager is initialized
-    private var isInitialized: Bool = false
-
-    /// Whether currently scanning for devices
-    var isScanning: Bool = false
+    // MARK: - Operation State
 
     /// Whether currently performing an operation (address, signing, etc.)
     var isOperating: Bool = false
 
-    /// List of discovered devices
-    var devices: [TrezorDeviceInfo] = []
-
-    /// Currently connected device
-    var connectedDevice: TrezorDeviceInfo?
-
-    /// Features of the connected device
-    var deviceFeatures: TrezorFeatures?
-
-    /// Device root fingerprint (hex string)
-    var deviceFingerprint: String?
-
-    /// Last error message
+    /// Last error message from a dev operation
     var error: String?
-
-    // MARK: - UI Dialog State
-
-    /// Show PIN entry dialog
-    var showPinEntry: Bool = false
-
-    /// Show passphrase entry dialog
-    var showPassphraseEntry: Bool = false
-
-    /// Show BLE pairing code dialog
-    var showPairingCode: Bool = false
-
-    /// Show "Confirm on device" overlay
-    var showConfirmOnDevice: Bool = false
-
-    /// Message for confirm on device overlay
-    var confirmMessage: String = ""
-
-    /// Show the "where to enter the passphrase" chooser (phone vs Trezor).
-    /// Only presented for devices that report on-device passphrase entry capability.
-    var showWalletModeChooser: Bool = false
-
-    // MARK: - Wallet Mode State
-
-    /// The currently selected wallet mode (standard / hidden-on-phone / hidden-on-device).
-    /// Drives the wallet-mode selector UI; the binding to the device session is applied
-    /// via setWalletMode (disconnect/reconnect).
-    var walletMode: TrezorWalletMode = .standard
-
-    /// Whether the connected device supports entering the passphrase on the Trezor itself.
-    var passphraseEntryCapable: Bool {
-        deviceFeatures?.passphraseEntryCapable == true
-    }
 
     // MARK: - Address Generation State
 
     /// Current derivation path
-    var derivationPath: String = "m/84'/\(defaultCoinTypeComponent)/0'/0/0"
+    var derivationPath: String = "m/84'/\(TrezorService.defaultCoinTypeComponent)/0'/0/0"
 
     /// Current script type for address generation
     var selectedScriptType: TrezorScriptType = .spendWitness
@@ -167,25 +98,10 @@ class TrezorViewModel {
     var messageToSign: String = "Hello, Trezor!"
 
     /// Path for message signing
-    var messageSigningPath: String = "m/84'/\(defaultCoinTypeComponent)/0'/0/0"
+    var messageSigningPath: String = "m/84'/\(TrezorService.defaultCoinTypeComponent)/0'/0/0"
 
     /// Signed message result
     var signedMessage: TrezorSignedMessageResponse?
-
-    // MARK: - Known Devices & Auto-Reconnect
-
-    /// Previously connected devices loaded from storage
-    var knownDevices: [TrezorKnownDevice] = []
-
-    /// Whether auto-reconnect is in progress
-    var isAutoReconnecting: Bool = false
-
-    /// Status text during auto-reconnect
-    var autoReconnectStatus: String?
-
-    /// Prevents a user-initiated disconnect from immediately reconnecting
-    /// when the disconnected device list appears.
-    private var suppressNextAutoReconnect = false
 
     // MARK: - Address Index
 
@@ -195,7 +111,7 @@ class TrezorViewModel {
     // MARK: - Public Key State
 
     /// Account-level derivation path for public key
-    var publicKeyPath: String = "m/84'/\(defaultCoinTypeComponent)/0'"
+    var publicKeyPath: String = "m/84'/\(TrezorService.defaultCoinTypeComponent)/0'"
 
     /// Retrieved xpub string
     var xpub: String?
@@ -356,57 +272,16 @@ class TrezorViewModel {
             !watcherEvents.isEmpty
     }
 
-    // MARK: - Bluetooth State
-
-    /// Current Bluetooth state — reads directly from BLEManager (@Observable chaining)
-    var bluetoothState: CBManagerState {
-        TrezorBLEManager.shared.bluetoothState
-    }
-
-    var isBridgeModeEnabled: Bool {
-        transport.isBridgeEnabled
-    }
-
     // MARK: - Private Properties
 
     private let trezorService = TrezorService.shared
     private let watcherService: TrezorWatcherServicing
-    private let transport = TrezorTransport.shared
-    private let uiHandler = TrezorUiHandler.shared
-    private var cancellables = Set<AnyCancellable>()
-    private var hasSetupSubscriptions = false
 
     // MARK: - Initialization
 
-    init(watcherService: TrezorWatcherServicing = TrezorService.shared) {
+    init(connection: TrezorManager, watcherService: TrezorWatcherServicing = TrezorService.shared) {
+        self.connection = connection
         self.watcherService = watcherService
-        selectedNetwork = Self.appDefaultCoinType
-        // Callback subscriptions are deferred to initialize() to avoid
-        // triggering BLE stack and Combine overhead at app launch.
-    }
-
-    /// Subscribe to callback publishers for UI notifications
-    private func setupCallbackSubscriptions() {
-        // Pairing code request
-        transport.needsPairingCodePublisher
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] in
-                self?.showPairingCode = true
-            }
-            .store(in: &cancellables)
-
-        // PIN request from device
-        uiHandler.needsPinPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] in
-                self?.showPinEntry = true
-            }
-            .store(in: &cancellables)
-
-        // Passphrase entry is now driven proactively by the wallet-mode selector
-        // (see setWalletMode / requestPassphraseWallet). The device callback
-        // `onPassphraseRequest` is answered silently from the selected mode, so there
-        // is no reactive passphrase prompt to subscribe to here.
     }
 
     // MARK: - Debug Log Helper
@@ -424,185 +299,12 @@ class TrezorViewModel {
         TrezorDebugLog.shared.log(message)
     }
 
-    // MARK: - State Reset Helpers
-
-    func clearWalletDerivedState() {
-        deviceFingerprint = nil
-        generatedAddress = nil
-        signedMessage = nil
-        xpub = nil
-        publicKeyHex = nil
-    }
-
-    func clearDisconnectedDeviceState(errorMessage: String? = nil) {
-        connectedDevice = nil
-        deviceFeatures = nil
-        clearWalletDerivedState()
-        error = errorMessage
-        showPinEntry = false
-        showPassphraseEntry = false
-        showConfirmOnDevice = false
-        showWalletModeChooser = false
-        uiHandler.setWalletMode(.standard)
-        walletMode = .standard
-    }
-
-    // MARK: - Manager Setup
-
-    /// Set up subscriptions and start BLE stack (synchronous, non-blocking).
-    /// Called from TrezorRootView's .task to prepare the UI layer.
-    func setup() {
-        guard !hasSetupSubscriptions else { return }
-        if !transport.isBridgeEnabled {
-            // Start BLE stack early so bluetoothState is updated by the time
-            // TrezorDeviceListView renders (the delegate callback fires async).
-            TrezorBLEManager.shared.ensureStarted()
-        }
-        setupCallbackSubscriptions()
-        hasSetupSubscriptions = true
-    }
-
-    /// Initialize the Trezor FFI manager (async, may be slow).
-    /// Called lazily before first scan/connect.
-    func initialize() async {
-        setup()
-
-        guard !isInitialized else { return }
-
-        do {
-            try await trezorService.initialize()
-            isInitialized = true
-            error = nil
-            trezorLog("TrezorViewModel initialized")
-        } catch {
-            self.error = errorMessage(from: error)
-            trezorLog("Failed to initialize Trezor: \(error)", level: "error")
-        }
-    }
-
-    // MARK: - Device Scanning
-
-    /// Start scanning for Trezor devices
-    /// - Parameter clearExisting: Whether to clear existing device list before scanning
-    func startScan(clearExisting: Bool = true) async {
-        if !isInitialized {
-            await initialize()
-        }
-
-        isScanning = true
-        error = nil
-
-        if clearExisting {
-            devices = []
-        }
-
-        if !transport.isBridgeEnabled {
-            // Start BLE scanning
-            transport.startBLEScanning()
-
-            // Wait for BLE to discover devices (like Android's 3-second scan)
-            // This ensures devices are found before we call the FFI enumerate
-            try? await Task.sleep(nanoseconds: 3_000_000_000) // 3 seconds
-
-            // Stop BLE scanning before calling FFI to prevent race conditions
-            transport.stopBLEScanning()
-        }
-
-        do {
-            // Trigger FFI scan which will use our transport callbacks
-            let foundDevices = try await trezorService.scan()
-
-            // Deduplicate by path (in case of duplicate scan results)
-            var seenPaths = Set<String>()
-            let uniqueDevices = foundDevices.filter { device in
-                if seenPaths.contains(device.path) {
-                    return false
-                }
-                seenPaths.insert(device.path)
-                return true
-            }
-
-            devices = uniqueDevices
-            trezorLog("Found \(uniqueDevices.count) Trezor devices (filtered from \(foundDevices.count))")
-        } catch {
-            self.error = errorMessage(from: error)
-            trezorLog("Scan failed: \(error)", level: "error")
-        }
-
-        isScanning = false
-    }
-
-    /// Stop scanning for devices
-    func stopScan() {
-        transport.stopBLEScanning()
-        isScanning = false
-    }
-
-    // MARK: - Connection
-
-    /// Connect to a device
-    func connect(device: TrezorDeviceInfo) async {
-        error = nil
-        suppressNextAutoReconnect = false
-
-        // Explicit user-initiated connect always opens the standard wallet — a
-        // passphrase/on-device selection left over from a previously connected device
-        // must not silently apply to a newly selected one.
-        uiHandler.setWalletMode(.standard)
-        walletMode = .standard
-
-        trezorLog("=== Connecting to device: \(device.path) ===")
-
-        do {
-            let features = try await trezorService.connect(deviceId: device.path, selection: uiHandler.currentSelection())
-            connectedDevice = device
-            deviceFeatures = features
-            showConfirmOnDevice = false
-
-            await saveCurrentDeviceAsKnown()
-            trezorLog("Connected to Trezor: \(device.path)")
-        } catch {
-            let errorMsg = errorMessage(from: error)
-            self.error = errorMsg
-            showConfirmOnDevice = false
-            trezorLog("Connection failed: \(error)", level: "error")
-        }
-    }
-
-    /// Disconnect from current device
-    func disconnect() async {
-        guard connectedDevice != nil else { return }
-        suppressNextAutoReconnect = true
-
-        // NOTE: the event watcher is intentionally NOT stopped here. It subscribes to
-        // Electrum directly and does not require a connected device, so it survives a
-        // disconnect and remains controllable from the device-list screen. It is only
-        // torn down on a network switch (different Electrum server) or via stopWatcher().
-
-        do {
-            try await trezorService.disconnect()
-            // Clear connection state but preserve device list for quick reconnection
-            clearDisconnectedDeviceState()
-
-            trezorLog("Disconnected from Trezor")
-        } catch {
-            // Even if disconnect fails, clear local state
-            clearDisconnectedDeviceState(errorMessage: errorMessage(from: error))
-            trezorLog("Disconnect failed: \(error)", level: "error")
-        }
-    }
-
-    /// Check if currently connected
-    var isConnected: Bool {
-        connectedDevice != nil
-    }
-
     // MARK: - Address Operations
 
     /// Get address from connected device
     /// - Parameter showOnDevice: Whether to display address on Trezor screen
     func getAddress(showOnDevice: Bool = true) async {
-        guard isConnected else {
+        guard connection.isConnected else {
             error = "Not connected to a Trezor"
             return
         }
@@ -613,19 +315,19 @@ class TrezorViewModel {
         do {
             let params = TrezorGetAddressParams(
                 path: derivationPath,
-                coin: selectedNetwork,
+                coin: connection.selectedNetwork,
                 showOnTrezor: showOnDevice,
                 scriptType: selectedScriptType
             )
 
             let response = try await trezorService.getAddress(params: params)
             generatedAddress = response.address
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
 
             trezorLog("Generated address: \(response.address)")
         } catch {
             self.error = errorMessage(from: error)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("Get address failed: \(error)", level: "error")
         }
 
@@ -636,7 +338,7 @@ class TrezorViewModel {
 
     /// Sign a message with the connected device
     func signMessage() async {
-        guard isConnected else {
+        guard connection.isConnected else {
             error = "Not connected to a Trezor"
             return
         }
@@ -653,17 +355,17 @@ class TrezorViewModel {
             let params = TrezorSignMessageParams(
                 path: messageSigningPath,
                 message: messageToSign,
-                coin: selectedNetwork
+                coin: connection.selectedNetwork
             )
 
             let response = try await trezorService.signMessage(params: params)
             signedMessage = response
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
 
             trezorLog("Message signed successfully")
         } catch {
             self.error = errorMessage(from: error)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("Sign message failed: \(error)", level: "error")
         }
 
@@ -672,7 +374,7 @@ class TrezorViewModel {
 
     /// Verify a signed message
     func verifyMessage(address: String, signature: String, message: String) async -> Bool {
-        guard isConnected else {
+        guard connection.isConnected else {
             error = "Not connected to a Trezor"
             return false
         }
@@ -685,11 +387,11 @@ class TrezorViewModel {
                 address: address,
                 signature: signature,
                 message: message,
-                coin: selectedNetwork
+                coin: connection.selectedNetwork
             )
 
             let isValid = try await trezorService.verifyMessage(params: params)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
 
             trezorLog("Message verification result: \(isValid)")
 
@@ -697,235 +399,12 @@ class TrezorViewModel {
             return isValid
         } catch {
             self.error = errorMessage(from: error)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("Verify message failed: \(error)", level: "error")
 
             isOperating = false
             return false
         }
-    }
-
-    // MARK: - UI Callbacks
-
-    /// Submit PIN from UI
-    func submitPin(_ pin: String) {
-        showPinEntry = false
-        uiHandler.submitPin(pin)
-    }
-
-    /// Cancel PIN entry
-    func cancelPin() {
-        showPinEntry = false
-        uiHandler.cancelPin()
-    }
-
-    /// Submit a host-entered passphrase from the UI — opens the corresponding hidden
-    /// wallet (or the standard wallet when empty) by resetting the session.
-    func submitPassphrase(_ passphrase: String) async {
-        showPassphraseEntry = false
-        showConfirmOnDevice = false
-        await setWalletMode(passphrase.isEmpty ? .standard : .passphraseHost, passphrase: passphrase)
-    }
-
-    /// Cancel passphrase entry
-    func cancelPassphrase() {
-        showPassphraseEntry = false
-        showConfirmOnDevice = false
-        showWalletModeChooser = false
-    }
-
-    // MARK: - Wallet Mode Selection
-
-    /// User tapped the "Standard" wallet option in the selector.
-    func selectStandardWallet() async {
-        guard walletMode != .standard else { return }
-        await setWalletMode(.standard)
-    }
-
-    /// User tapped the "Passphrase" wallet option. On a capable device this offers a
-    /// choice of where to enter the passphrase; otherwise it goes straight to host entry.
-    func requestPassphraseWallet() {
-        if passphraseEntryCapable {
-            showWalletModeChooser = true
-        } else {
-            showPassphraseEntry = true
-        }
-    }
-
-    /// Wallet-mode chooser: user chose to enter the passphrase on this phone.
-    func choosePhonePassphraseEntry() {
-        showWalletModeChooser = false
-        showPassphraseEntry = true
-    }
-
-    /// Wallet-mode chooser: user chose to enter the passphrase on the Trezor.
-    func chooseDevicePassphraseEntry() async {
-        showWalletModeChooser = false
-        await setWalletMode(.passphraseDevice)
-    }
-
-    /// Switch between wallet modes. The Trezor caches the passphrase for the whole
-    /// session, so switching requires a fresh session: this records the desired mode,
-    /// then disconnects and reconnects by path. Mirrors bitkit-android's setWalletMode.
-    func setWalletMode(_ mode: TrezorWalletMode, passphrase: String = "") async {
-        guard let device = connectedDevice else {
-            error = "Not connected to a Trezor"
-            return
-        }
-
-        isOperating = true
-        error = nil
-        trezorLog("=== Switching wallet mode to \(mode); resetting session ===")
-
-        // Reset the session. We call the service directly (not the VM's disconnect())
-        // so connectedDevice/deviceFeatures stay populated for the reconnect.
-        do {
-            try await trezorService.disconnect()
-        } catch {
-            trezorLog("Disconnect before wallet-mode switch failed: \(error)", level: "warn")
-        }
-
-        // Results derived from the previous wallet are no longer valid once the
-        // session has been reset for a different wallet mode.
-        clearWalletDerivedState()
-
-        // Brief settle delay before reconnecting (matches Android's reconnect delay).
-        try? await Task.sleep(nanoseconds: 300_000_000)
-
-        // Record the selection AFTER the disconnect so it survives into the new session.
-        // THP reads it via currentSelection() to bind the passphrase at session creation;
-        // non-THP devices re-request it mid-operation and are answered from the same value.
-        uiHandler.setWalletMode(mode, hostPassphrase: passphrase)
-        walletMode = mode
-
-        do {
-            let features = try await trezorService.connect(deviceId: device.path, selection: uiHandler.currentSelection())
-            connectedDevice = device
-            deviceFeatures = features
-            showConfirmOnDevice = false
-            trezorLog("Reconnected with wallet mode \(mode)")
-        } catch {
-            clearDisconnectedDeviceState(errorMessage: errorMessage(from: error))
-            trezorLog("Reconnect after wallet-mode switch failed: \(error)", level: "error")
-        }
-
-        isOperating = false
-    }
-
-    /// Submit pairing code from UI
-    func submitPairingCode(_ code: String) {
-        showPairingCode = false
-        transport.submitPairingCode(code)
-    }
-
-    /// Cancel pairing code entry
-    func cancelPairingCode() {
-        showPairingCode = false
-        transport.cancelPairingCode()
-    }
-
-    /// Dismiss confirm on device overlay
-    func dismissConfirmOnDevice() {
-        showConfirmOnDevice = false
-        confirmMessage = ""
-    }
-
-    // MARK: - Known Devices
-
-    /// Load known devices from storage
-    func loadKnownDevices() {
-        knownDevices = TrezorKnownDeviceStorage.loadAll()
-    }
-
-    /// Save the currently connected device as a known device, capturing its account
-    /// xpubs so watch-only balances/activity stay available while disconnected.
-    func saveCurrentDeviceAsKnown() async {
-        guard let device = connectedDevice else { return }
-        let previous = TrezorKnownDeviceStorage.loadAll().first { $0.id == device.id }
-        let fetched = await fetchAccountXpubs()
-        let mergedXpubs = (previous?.xpubs ?? [:]).merging(fetched) { _, new in new }
-        let known = TrezorKnownDevice(
-            id: device.id,
-            name: device.name ?? "Trezor",
-            path: device.path,
-            transportType: device.transportType == .bluetooth ? "bluetooth" : "usb",
-            label: device.label ?? deviceFeatures?.label,
-            model: device.model ?? deviceFeatures?.model,
-            lastConnectedAt: Date(),
-            xpubs: mergedXpubs
-        )
-        TrezorKnownDeviceStorage.save(known)
-        loadKnownDevices()
-        trezorLog("Saved known device: \(known.name) with \(mergedXpubs.count) xpubs")
-    }
-
-    /// Read the account-level xpub for every address type from the connected device.
-    /// Per-type failures are swallowed so a single missing type doesn't block the rest.
-    func fetchAccountXpubs() async -> [String: String] {
-        var result: [String: String] = [:]
-        for addressType in HwAddressType.allCases {
-            do {
-                let params = TrezorGetPublicKeyParams(
-                    path: addressType.accountDerivationPath(network: selectedNetwork),
-                    coin: selectedNetwork,
-                    showOnTrezor: false
-                )
-                let response = try await trezorService.getPublicKey(params: params)
-                result[addressType.settingsString] = response.xpub
-            } catch {
-                trezorLog("Could not read xpub for '\(addressType.settingsString)': \(error)", level: "warn")
-            }
-        }
-        return result
-    }
-
-    /// Forget a known device — removes from storage and clears credentials
-    func forgetDevice(id: String) async {
-        // Find the device to get its path for credential clearing
-        if let device = knownDevices.first(where: { $0.id == id }) {
-            do {
-                try await trezorService.clearCredentials(deviceId: device.path)
-            } catch {
-                trezorLog("Failed to clear credentials for forgotten device: \(error)", level: "warn")
-            }
-            TrezorCredentialStorage.delete(deviceId: device.path)
-        }
-        TrezorKnownDeviceStorage.remove(id: id)
-        loadKnownDevices()
-        trezorLog("Forgot device: \(id)")
-    }
-
-    // MARK: - Auto-Reconnect
-
-    /// Automatically scan and reconnect to the first matching known device
-    func autoReconnect() async {
-        guard !knownDevices.isEmpty else { return }
-        guard !isAutoReconnecting else { return }
-        if suppressNextAutoReconnect {
-            suppressNextAutoReconnect = false
-            trezorLog("Auto-reconnect: skipped after manual disconnect")
-            return
-        }
-
-        isAutoReconnecting = true
-        autoReconnectStatus = "Scanning for known devices..."
-        trezorLog("Auto-reconnect: starting scan")
-
-        await startScan(clearExisting: true)
-
-        // Find the first scanned device that matches a known device
-        let knownIds = Set(knownDevices.map(\.id))
-        if let match = devices.first(where: { knownIds.contains($0.id) }) {
-            autoReconnectStatus = "Connecting to \(match.label ?? match.name ?? "Trezor")..."
-            trezorLog("Auto-reconnect: found known device \(match.path)")
-            await connect(device: match)
-        } else {
-            autoReconnectStatus = nil
-            trezorLog("Auto-reconnect: no known devices found nearby")
-        }
-
-        isAutoReconnecting = false
-        autoReconnectStatus = nil
     }
 
     // MARK: - Address Index
@@ -955,7 +434,7 @@ class TrezorViewModel {
 
     /// Get public key (xpub) from connected device
     func getPublicKey(showOnDevice: Bool = false) async {
-        guard isConnected else {
+        guard connection.isConnected else {
             error = "Not connected to a Trezor"
             return
         }
@@ -966,19 +445,19 @@ class TrezorViewModel {
         do {
             let params = TrezorGetPublicKeyParams(
                 path: publicKeyPath,
-                coin: selectedNetwork,
+                coin: connection.selectedNetwork,
                 showOnTrezor: showOnDevice
             )
 
             let response = try await trezorService.getPublicKey(params: params)
             xpub = response.xpub
             publicKeyHex = response.publicKey
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
 
             trezorLog("Got public key for path: \(publicKeyPath)")
         } catch {
             self.error = errorMessage(from: error)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("Get public key failed: \(error)", level: "error")
         }
 
@@ -989,7 +468,7 @@ class TrezorViewModel {
 
     /// Sign a Bitcoin transaction
     func signTx(params: TrezorSignTxParams) async -> TrezorSignedTx? {
-        guard isConnected else {
+        guard connection.isConnected else {
             error = "Not connected to a Trezor"
             return nil
         }
@@ -999,13 +478,13 @@ class TrezorViewModel {
 
         do {
             let result = try await trezorService.signTx(params: params)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("Transaction signed successfully")
             isOperating = false
             return result
         } catch {
             self.error = errorMessage(from: error)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("Sign tx failed: \(error)", level: "error")
             isOperating = false
             return nil
@@ -1018,7 +497,7 @@ class TrezorViewModel {
     /// - Parameter psbtBase64: Base64-encoded PSBT data
     /// - Returns: The signed transaction, or nil on failure
     func signTxFromPsbt(psbtBase64: String) async -> TrezorSignedTx? {
-        guard isConnected else {
+        guard connection.isConnected else {
             error = "Not connected to a Trezor"
             return nil
         }
@@ -1029,15 +508,15 @@ class TrezorViewModel {
         do {
             let result = try await trezorService.signTxFromPsbt(
                 psbtBase64: psbtBase64,
-                network: selectedNetwork
+                network: connection.selectedNetwork
             )
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("PSBT signed successfully")
             isOperating = false
             return result
         } catch {
             self.error = errorMessage(from: error)
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("Sign PSBT failed: \(error)", level: "error")
             isOperating = false
             return nil
@@ -1048,7 +527,7 @@ class TrezorViewModel {
 
     /// Get the device's master root fingerprint
     func getDeviceFingerprint() async {
-        guard isConnected else {
+        guard connection.isConnected else {
             error = "Not connected to a Trezor"
             return
         }
@@ -1058,7 +537,7 @@ class TrezorViewModel {
 
         do {
             let fingerprint = try await trezorService.getDeviceFingerprint()
-            deviceFingerprint = fingerprint
+            connection.deviceFingerprint = fingerprint
             trezorLog("Device fingerprint: \(fingerprint)")
         } catch {
             self.error = errorMessage(from: error)
@@ -1068,45 +547,18 @@ class TrezorViewModel {
         isOperating = false
     }
 
-    // MARK: - Electrum URL Helpers
-
-    /// Get the Electrum server URL for a specific network (hardcoded per-network URLs)
-    static func electrumUrlForNetwork(_ network: TrezorCoinType) -> String {
-        if network == .regtest, let trezorElectrumUrl = Env.trezorElectrumUrl {
-            return trezorElectrumUrl
-        }
-
-        switch network {
-        case .bitcoin:
-            return "ssl://bitkit.to:9999"
-        case .testnet, .signet:
-            return "ssl://electrum.blockstream.info:60002"
-        case .regtest:
-            return "ssl://electrs.bitkit.stag0.blocktank.to:9999"
-        }
-    }
-
-    /// Get the current Electrum server URL from configuration (uses app's configured server)
-    static func getElectrumUrl() -> String {
-        let configService = ElectrumConfigService()
-        let server = configService.getCurrentServer()
-        return server.fullUrl.isEmpty ? Env.electrumServerUrl : server.fullUrl
-    }
-
     // MARK: - Network Switching
 
-    /// Switch the dashboard's network independently of the app's global network
-    func setSelectedNetwork(_ network: TrezorCoinType) {
-        guard network != selectedNetwork else { return }
-        selectedNetwork = network
-
+    /// React to a dashboard network switch by resetting the dev-tool derivation paths,
+    /// results and the dev watcher. The network change itself is owned by `TrezorManager`.
+    func handleNetworkChange() {
         // A running watcher is bound to the previous network's Electrum server.
         stopWatcher()
 
         // Reset derivation paths with the new coin type
-        derivationPath = "m/84'/\(coinTypeComponent)/0'/0/0"
-        publicKeyPath = "m/84'/\(coinTypeComponent)/0'"
-        messageSigningPath = "m/84'/\(coinTypeComponent)/0'/0/0"
+        derivationPath = "m/84'/\(connection.coinTypeComponent)/0'/0/0"
+        publicKeyPath = "m/84'/\(connection.coinTypeComponent)/0'"
+        messageSigningPath = "m/84'/\(connection.coinTypeComponent)/0'/0/0"
         addressIndex = 0
 
         // Clear results from previous network
@@ -1119,8 +571,6 @@ class TrezorViewModel {
         addressResult = nil
         lookupError = nil
         resetSendFlow()
-
-        trezorLog("Switched dashboard network to \(network)")
     }
 
     // MARK: - Balance Lookup Operations
@@ -1161,7 +611,7 @@ class TrezorViewModel {
         addressResult = nil
         resetSendFlow()
 
-        let electrumUrl = Self.electrumUrlForNetwork(selectedNetwork)
+        let electrumUrl = TrezorService.electrumUrlForNetwork(connection.selectedNetwork)
 
         do {
             switch Self.detectInputType(trimmedInput) {
@@ -1169,14 +619,14 @@ class TrezorViewModel {
                 accountResult = try await trezorService.getAccountInfo(
                     extendedKey: trimmedInput,
                     electrumUrl: electrumUrl,
-                    network: selectedNetwork,
+                    network: connection.selectedNetwork,
                     scriptType: onchainAccountTypeSelection.accountType
                 )
             case .address:
                 addressResult = try await trezorService.getAddressInfo(
                     address: trimmedInput,
                     electrumUrl: electrumUrl,
-                    network: selectedNetwork
+                    network: connection.selectedNetwork
                 )
             case .unknown:
                 lookupError = "Unrecognized input. Enter a Bitcoin address or extended public key (xpub/ypub/zpub/tpub/upub/vpub)."
@@ -1233,13 +683,13 @@ class TrezorViewModel {
         txHistoryError = nil
         txHistoryResult = nil
 
-        let electrumUrl = Self.electrumUrlForNetwork(selectedNetwork)
+        let electrumUrl = TrezorService.electrumUrlForNetwork(connection.selectedNetwork)
 
         do {
             txHistoryResult = try await trezorService.getTransactionHistory(
                 extendedKey: trimmedKey,
                 electrumUrl: electrumUrl,
-                network: selectedNetwork,
+                network: connection.selectedNetwork,
                 scriptType: onchainAccountTypeSelection.accountType
             )
         } catch {
@@ -1261,14 +711,14 @@ class TrezorViewModel {
         txDetailError = nil
         txDetailResult = nil
 
-        let electrumUrl = Self.electrumUrlForNetwork(selectedNetwork)
+        let electrumUrl = TrezorService.electrumUrlForNetwork(connection.selectedNetwork)
 
         do {
             txDetailResult = try await trezorService.getTransactionDetail(
                 extendedKey: trimmedKey,
                 electrumUrl: electrumUrl,
                 txid: trimmedTxid,
-                network: selectedNetwork,
+                network: connection.selectedNetwork,
                 scriptType: onchainAccountTypeSelection.accountType
             )
         } catch {
@@ -1318,9 +768,9 @@ class TrezorViewModel {
         // Ensure we have the device fingerprint for proper PSBT derivation paths.
         // Without it, BDK produces relative paths (e.g. m/0/0) that the Trezor
         // rejects as "Forbidden key path".
-        if deviceFingerprint == nil {
+        if connection.deviceFingerprint == nil {
             do {
-                deviceFingerprint = try await trezorService.getDeviceFingerprint()
+                connection.deviceFingerprint = try await trezorService.getDeviceFingerprint()
             } catch {
                 trezorLog("Failed to get device fingerprint: \(error)", level: "error")
                 sendError = "Failed to get device fingerprint"
@@ -1338,13 +788,13 @@ class TrezorViewModel {
             ? .sendMax(address: address)
             : .payment(address: address, amountSats: UInt64(sendAmountSats) ?? 0)
 
-        let electrumUrl = Self.electrumUrlForNetwork(selectedNetwork)
-        let network = toNetwork(selectedNetwork)
+        let electrumUrl = TrezorService.electrumUrlForNetwork(connection.selectedNetwork)
+        let network = toNetwork(connection.selectedNetwork)
 
         let wallet = WalletParams(
             extendedKey: extendedKey,
             electrumUrl: electrumUrl,
-            fingerprint: deviceFingerprint,
+            fingerprint: connection.deviceFingerprint,
             network: network,
             accountType: accountInfo.accountType
         )
@@ -1413,7 +863,7 @@ class TrezorViewModel {
             return
         }
 
-        guard isConnected else {
+        guard connection.isConnected else {
             sendError = "Not connected to a Trezor"
             return
         }
@@ -1429,9 +879,9 @@ class TrezorViewModel {
             trezorLog("Calling trezor signTxFromPsbt...")
             let signedTx = try await trezorService.signTxFromPsbt(
                 psbtBase64: psbt,
-                network: selectedNetwork
+                network: connection.selectedNetwork
             )
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
 
             trezorLog("=== signComposedTx SUCCESS ===")
             trezorLog("signatures=\(signedTx.signatures.count), txid=\(signedTx.txid ?? "nil"), rawTxLen=\(signedTx.serializedTx.count)")
@@ -1439,7 +889,7 @@ class TrezorViewModel {
             signedTxResult = signedTx
             sendStep = .signed
         } catch {
-            showConfirmOnDevice = false
+            connection.showConfirmOnDevice = false
             trezorLog("signComposedTx FAILED: \(error)", level: "error")
             sendError = errorMessage(from: error)
         }
@@ -1454,7 +904,7 @@ class TrezorViewModel {
         isBroadcasting = true
         sendError = nil
 
-        let electrumUrl = Self.electrumUrlForNetwork(selectedNetwork)
+        let electrumUrl = TrezorService.electrumUrlForNetwork(connection.selectedNetwork)
 
         do {
             let txid = try await trezorService.broadcastRawTx(serializedTx: rawTx, electrumUrl: electrumUrl)
@@ -1583,24 +1033,6 @@ class TrezorViewModel {
         return cleanedMessage
     }
 
-    // MARK: - Credential Management
-
-    /// Clear stored credentials for current device
-    func clearCredentials() async {
-        guard let device = connectedDevice else {
-            error = "No device connected"
-            return
-        }
-
-        do {
-            try await trezorService.clearCredentials(deviceId: device.path)
-            trezorLog("Cleared credentials for \(device.path)")
-        } catch {
-            self.error = errorMessage(from: error)
-            trezorLog("Failed to clear credentials: \(error)", level: "error")
-        }
-    }
-
     // MARK: - Event Watcher Operations
 
     /// Copy the most recently retrieved xpub into the watcher's extended-key field.
@@ -1625,13 +1057,13 @@ class TrezorViewModel {
         }
 
         let watcherId = UUID().uuidString
-        let network = selectedNetwork
+        let network = connection.selectedNetwork
         let accountType = onchainAccountTypeSelection.accountType
 
         let params = WatcherParams(
             watcherId: watcherId,
             extendedKey: key,
-            electrumUrl: Self.electrumUrlForNetwork(network),
+            electrumUrl: TrezorService.electrumUrlForNetwork(network),
             network: toNetwork(network),
             accountType: accountType,
             gapLimit: gapLimit
@@ -1662,7 +1094,7 @@ class TrezorViewModel {
                 return
             }
 
-            if selectedNetwork != network {
+            if connection.selectedNetwork != network {
                 try? watcherService.stopWatcher(watcherId: watcherId)
                 finishStoppedWatcherStartup(watcherId: watcherId)
                 return
@@ -1852,22 +1284,22 @@ class TrezorViewModel {
 
     func testShowPinPrompt() {
         guard Env.isTrezorEmulatorTesting else { return }
-        showPinEntry = true
+        connection.showPinEntry = true
     }
 
     func testShowPassphrasePrompt() {
         guard Env.isTrezorEmulatorTesting else { return }
-        showPassphraseEntry = true
+        connection.showPassphraseEntry = true
     }
 
     func testShowPairingCodePrompt() {
         guard Env.isTrezorEmulatorTesting else { return }
-        showPairingCode = true
+        connection.showPairingCode = true
     }
 
     func testShowConfirmOnDevicePrompt() {
         guard Env.isTrezorEmulatorTesting else { return }
-        confirmMessage = "Confirm test action on your Trezor"
-        showConfirmOnDevice = true
+        connection.confirmMessage = "Confirm test action on your Trezor"
+        connection.showConfirmOnDevice = true
     }
 }
