+++ Bitkit/Views/Trezor/TrezorBalanceLookupView.swift
@@ -56,6 +56,7 @@ private struct LookupButtonWrapper: View {
 /// keeping the parent body free of ViewModel property accesses.
 private struct BalanceLookupResultsSection: View {
     let input: String
+    @Environment(TrezorManager.self) private var trezorManager
     @Environment(TrezorViewModel.self) private var trezor
 
     private var hasResults: Bool {
@@ -89,7 +90,7 @@ private struct BalanceLookupResultsSection: View {
                     isComposing: trezor.isComposing,
                     isOperating: trezor.isOperating,
                     isBroadcasting: trezor.isBroadcasting,
-                    isDeviceConnected: trezor.isConnected,
+                    isDeviceConnected: trezorManager.isConnected,
                     composeResult: trezor.composeResult,
                     signedTxResult: trezor.signedTxResult,
                     broadcastTxid: trezor.broadcastTxid,
